backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
python-dotenv==1.0.0
psycopg2-binary==2.9.10
PyJWT==2.8.0
Flask-CORS
celery==5.5.3
redis==6.4.0
pypdf==6.1.1
langchain-groq==0.2.1
langchain-core==0.3.26
python-telegram-bot==21.7
Flask-Mail==0.10.0
langchain==0.3.13
langchain-community
langchain-huggingface
langchain-pinecone
faiss-cpu
pinecone-client
slack-sdk
Flask-JWT-Extended